+++ src/components/private-route/private-route.tsx
@@ -1,7 +1,7 @@
 import {Navigate} from 'react-router-dom';
 import {AppRoute, AuthorizationStatus} from '../../const';
 import {useAppSelector} from '../../hooks';
-import { selectAuthorizationStatus } from '../../store/selectors';
+import {selectAuthorizationStatus} from '../../store/user-process/selectors';
 
 type PrivateRouteProps = {
   children: JSX.Element;
