+++ src/pages/main-page/main-page.tsx
@@ -1,17 +1,13 @@
 import {Helmet} from 'react-helmet-async';
-import {Offer} from '../../types';
 import Header from '../../components/header/header';
 import Cities from '../../components/cities/cities';
-import SortForm from '../../components/sort-form/sort-form';
-import CardsList from '../../components/cards-list/cards-list';
-import Map from '../../components/map/map';
 import MainEmpty from '../../components/main-empty/main-empty';
 import cn from 'classnames';
-import {useState} from 'react';
 import {useAppSelector} from '../../hooks';
-import {getMapPoints} from '../../utils';
-import {selectActiveCity, selectCitySortOffers, selectIsErrorInOffersDataLoading} from '../../store/selectors';
 import ServerErrorPage from '../server-error-page/server-error-page';
+import MainPageInner from './main-page-inner';
+import {Page} from '../../const';
+import {selectCitySortOffers, selectIsErrorInOffersDataLoading} from '../../store/cards-process/selectors';
 
 type MainPageProps = {
   cities: string[];
@@ -19,38 +15,11 @@ type MainPageProps = {
 
 function MainPage({cities}: MainPageProps): JSX.Element {
   const isErrorInOffersDataLoading = useAppSelector(selectIsErrorInOffersDataLoading);
-  const [activeCard, setActiveCard] = useState<Offer | null>(null);
-  const activeCity = useAppSelector(selectActiveCity);
   const sortedOffers = useAppSelector(selectCitySortOffers);
-  const mapPoints = getMapPoints(sortedOffers);
 
   if (isErrorInOffersDataLoading) {
-    return <ServerErrorPage/>;
+    return <ServerErrorPage page={Page.Cities}/>;
   }
-
-  const handleCardHover = (offer: Offer | null) => {
-    setActiveCard(offer);
-  };
-
-  const mainInner = (sortedOffers.length ? (
-    <>
-      <section className='cities__places places'>
-        <h2 className='visually-hidden'>Places</h2>
-        <b className='places__found'>{sortedOffers.length} places to stay in {activeCity}</b>
-        <SortForm />
-        <CardsList
-          offers={sortedOffers}
-          onCardHover={handleCardHover}
-        />
-      </section>
-      <div className='cities__right-section'>
-        <Map page='cities' offers={mapPoints} selectedOffer={activeCard}/>
-      </div>
-    </>
-  )
-    :
-    <MainEmpty/>
-  );
 
   return (
     <div className='page page--gray page--main'>
@@ -62,7 +31,7 @@ function MainPage({cities}: MainPageProps): JSX.Element {
         <Cities cities={cities}/>
         <div className='cities'>
           <div className={cn('cities__places-container', 'container', {'cities__places-container--empty':!sortedOffers.length})}>
-            {mainInner}
+            {sortedOffers.length ? <MainPageInner sortedOffers={sortedOffers}/> : <MainEmpty/>}
           </div>
         </div>
       </main>
