+++ src/store/index.ts
@@ -1,16 +1,15 @@
 import {configureStore} from '@reduxjs/toolkit';
-import {reducer} from './reducer';
 import {createAPI} from '../services/api';
-import { redirect } from './middlewares/redirect';
+import {rootReducer} from './root-reducer';
 
 export const api = createAPI();
 
 export const store = configureStore({
-  reducer,
+  reducer: rootReducer,
   middleware: (getDefaultMiddleware) =>
     getDefaultMiddleware({
       thunk: {
         extraArgument: api,
       },
-    }).concat(redirect),
+    }),
 });
