--- src/store/middlewares/redirect.tsx
@@ -1,17 +1,0 @@
-import {PayloadAction} from '@reduxjs/toolkit';
-import browserHistory from '../../browser-history';
-import {Middleware} from 'redux';
-import {reducer} from '../reducer';
-
-type Reducer = ReturnType<typeof reducer>;
-
-export const redirect: Middleware<unknown, Reducer> =
-  () =>
-    (next) =>
-      (action: PayloadAction<string>) => {
-        if (action.type === 'redirectToRoute') {
-          browserHistory.push(action.payload);
-        }
-
-        return next(action);
-      };
