+++ src/store/user-process/selectors.ts
@@ -1,0 +1,6 @@
+import {NameSpace} from '../../const';
+import {AuthorizationStatus} from '../../const';
+import {AppState} from '../../types';
+
+export const selectAuthorizationStatus = (state: AppState): AuthorizationStatus => state[NameSpace.User].authorizationStatus;
+export const selectUserName = (state: AppState): string => state[NameSpace.User].userInfo && state[NameSpace.User].userInfo.email;
